+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `WithChildren` trait.
+- `HtmlElement`: Added `text_content_mut` method.
 
 ## [0.2.0] - 2024-01-06